refactor: consolidate tooltip components
